refactor(test): consistent package structure for tests
