+++ pkg/parse_hal/hal_to_echo_test.go
@@ -13,143 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package parse_hal
+package parse_hal_test
 
 import (
+	"bytes"
+	"io/ioutil"
+	"path/filepath"
 	"reflect"
 	"testing"
 
+	"github.com/spinnaker/kleat/internal/protoyaml"
+
+	"github.com/spinnaker/kleat/pkg/parse_hal"
+
 	"github.com/spinnaker/kleat/api/client"
-	"github.com/spinnaker/kleat/api/client/artifact"
 	"github.com/spinnaker/kleat/api/client/ci"
-	"github.com/spinnaker/kleat/api/client/cloudprovider"
 	"github.com/spinnaker/kleat/api/client/config"
 	"github.com/spinnaker/kleat/api/client/notification"
 	"github.com/spinnaker/kleat/api/client/pubsub"
 )
-
-var halToClouddriverTests = []struct {
-	n     string
-	h     *config.Hal
-	wantC *config.Clouddriver
-}{
-	{
-		"Empty hal config",
-		&config.Hal{},
-		&config.Clouddriver{},
-	},
-	{
-		"Empty providers",
-		&config.Hal{
-			Providers: &cloudprovider.Providers{},
-		},
-		&config.Clouddriver{},
-	},
-	{
-		"Empty Kubernetes provider",
-		&config.Hal{
-			Providers: &cloudprovider.Providers{
-				Kubernetes: &cloudprovider.Kubernetes{},
-			},
-		},
-		&config.Clouddriver{
-			Kubernetes: &cloudprovider.Kubernetes{},
-		},
-	},
-	{
-		"Kubernetes account",
-		&config.Hal{
-			Providers: &cloudprovider.Providers{
-				Kubernetes: &cloudprovider.Kubernetes{
-					Enabled: true,
-					Accounts: []*cloudprovider.KubernetesAccount{
-						{
-							Name:           "my-account",
-							Kinds:          []string{"deployment"},
-							OmitNamespaces: []string{"kube-system"},
-						},
-					},
-					PrimaryAccount: "my-account",
-				},
-			},
-		},
-		&config.Clouddriver{
-			Kubernetes: &cloudprovider.Kubernetes{
-				Enabled: true,
-				Accounts: []*cloudprovider.KubernetesAccount{
-					{
-						Name:           "my-account",
-						Kinds:          []string{"deployment"},
-						OmitNamespaces: []string{"kube-system"},
-					},
-				},
-				PrimaryAccount: "my-account",
-			},
-		},
-	},
-	{
-		"Empty artifacts",
-		&config.Hal{
-			Artifacts: &artifact.Artifacts{},
-		},
-		&config.Clouddriver{
-			Artifacts: &artifact.Artifacts{},
-		},
-	},
-	{
-		"Empty GCS artifacts",
-		&config.Hal{
-			Artifacts: &artifact.Artifacts{
-				Gcs: &artifact.Gcs{},
-			},
-		},
-		&config.Clouddriver{
-			Artifacts: &artifact.Artifacts{
-				Gcs: &artifact.Gcs{},
-			},
-		},
-	},
-	{
-		"GCS artifact account",
-		&config.Hal{
-			Artifacts: &artifact.Artifacts{
-				Gcs: &artifact.Gcs{
-					Enabled: true,
-					Accounts: []*artifact.GcsAccount{
-						{
-							Name:     "my-account",
-							JsonPath: "/var/secrets/my-key.json",
-						},
-					},
-				},
-			},
-		},
-		&config.Clouddriver{
-			Artifacts: &artifact.Artifacts{
-				Gcs: &artifact.Gcs{
-					Enabled: true,
-					Accounts: []*artifact.GcsAccount{
-						{
-							Name:     "my-account",
-							JsonPath: "/var/secrets/my-key.json",
-						},
-					},
-				},
-			},
-		},
-	},
-}
-
-func TestHalToClouddriver(t *testing.T) {
-	for _, tt := range halToClouddriverTests {
-		t.Run(tt.n, func(t *testing.T) {
-			gotC := HalToClouddriver(tt.h)
-			if !reflect.DeepEqual(gotC, tt.wantC) {
-				t.Errorf("Expected hal config to generate %v, got %v", tt.wantC, gotC)
-			}
-		})
-	}
-}
 
 var halToEchoTests = []struct {
 	n     string
@@ -337,10 +219,51 @@ var halToEchoTests = []struct {
 func TestHalToEcho(t *testing.T) {
 	for _, tt := range halToEchoTests {
 		t.Run(tt.n, func(t *testing.T) {
-			gotC := HalToEcho(tt.h)
+			gotC := parse_hal.HalToEcho(tt.h)
 			if !reflect.DeepEqual(gotC, tt.wantE) {
 				t.Errorf("Expected hal config to generate %v, got %v", tt.wantE, gotC)
 			}
 		})
 	}
+}
+
+func TestHalToEchoYaml(t *testing.T) {
+	h, err := parse_hal.ParseHalConfig(filepath.Join("../../testdata", "halconfig.yml"))
+	if err != nil {
+		t.Errorf(err.Error())
+	}
+
+	gotE := parse_hal.HalToEcho(h)
+
+	wantE, err := parseEchoConfig(filepath.Join("../../testdata", "echo.yml"))
+	if err != nil {
+		t.Errorf(err.Error())
+	}
+
+	want, err := protoyaml.Marshal(wantE)
+	if err != nil {
+		t.Errorf(err.Error())
+	}
+
+	got, err := protoyaml.Marshal(gotE)
+	if err != nil {
+		t.Errorf(err.Error())
+	}
+
+	res := bytes.Compare(want, got)
+	if res != 0 {
+		t.Errorf("Expected generated Echo config to match contents of echo.yml, but got:\n" + string(got))
+	}
+}
+
+func parseEchoConfig(fn string) (*config.Echo, error) {
+	dat, err := ioutil.ReadFile(fn)
+
+	h := config.Echo{}
+	err = protoyaml.UnmarshalStrict([]byte(dat), &h)
+	if err != nil {
+		return nil, err
+	}
+
+	return &h, nil
 }
